+++ examples/Example120_ThreeRegions1D.jl
@@ -8,11 +8,14 @@ using VoronoiFVM
 if installed("Plots")
     using Plots
 end
-
+ 
 
 
 
 function main(;n=30,doplot=false,verbose=false,dense=false)
+    if !installed("Plots")
+        doplot=false
+    end
     h=3.0/(n-1)
     X=collect(0:h:3.0)
     grid=VoronoiFVM.Grid(X)
@@ -26,7 +29,7 @@ function main(;n=30,doplot=false,verbose=false,dense=false)
     
     if doplot
         p=Plots.plot()
-        fvmplot!(p,grid)
+        VoronoiFVM.plot(Plots,grid,p=p)
         gui(p)
         readline()
     end
@@ -109,9 +112,9 @@ function main(;n=30,doplot=false,verbose=false,dense=false)
         testval=U[2,15]
         if doplot
             p=Plots.plot()
-            fvmplot!(p,subgrid1, U[1,:],label="spec1", color=(0.5,0,0))
-            fvmplot!(p,subgrid2, U[2,:],label="spec2", color=(0.0,0.5,0))
-            fvmplot!(p,subgrid3, U[3,:],label="spec3", color=(0.0,0.0,0.5))
+            VoronoiFVM.plot(Plots,subgrid1, U[1,:],label="spec1", color=(0.5,0,0),p=p,show=false)
+            VoronoiFVM.plot(Plots,subgrid2, U[2,:],label="spec2", color=(0.0,0.5,0),p=p,show=false)
+            VoronoiFVM.plot(Plots,subgrid3, U[3,:],label="spec3", color=(0.0,0.0,0.5),p=p,show=false)
             gui(p)
         end
     end