+++ examples/Example125_TestFunctions1D.jl
@@ -11,6 +11,10 @@ if installed("Plots")
 end
 
 function main(;n=100,doplot=false,verbose=false,dense=false)
+    if !installed("Plots")
+        doplot=false
+    end
+
     h=1/n
     grid=VoronoiFVM.Grid(collect(0:h:1))