+++ examples/Example220_NonlinearPoisson2D_BoundarySpecies.jl
@@ -10,6 +10,9 @@ if installed("Plots")
 end
 
 function main(;n=10,doplot=false,verbose=false,dense=false)
+    if !installed("Plots")
+        doplot=false
+    end
     
     
     h=1.0/convert(Float64,n)